backend/exam_portal/requirements.txt
Django==4.2.29
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv
opencv-python
numpy
Pillow
